# src/main/java/com/p1/db/p1_db/entity/User.java
package com.p1.db.p1_db.entity;

import javax.persistence.Entity;
import com.p1.db.p1_db.entity.base.UserBase;

@Entity(name="user_app")
public class User extends UserBase {

    public User() {}

    public User(String id) {
        this.set_id(Long.valueOf(id));
    }

	//OVERRIDE HERE YOUR CUSTOM MAPPER
	
	
}